Changes to output, build system and console bug fixes
Fixed a bug that prevented the first command to be called prior to the
Subliminol
Console instatiation. This was due to an order of operations issue.
The active view was retrieved after the console was created, making the
console
the command source. Since it was just created and had no selection there
would
be no command to run.

The settings object is now passed from the SubliminolCommand object to
the
individual SubliminolCallBase subclasses. This may need to change since
these
objects are called from another thread.

Added sbnl_log() as a means of managing output more consistency. All
reporting
is now handled with this function. A level kwarg can be passed into it
to alow
a priority level to be assigned to a particular log message. This alows
debug
logging calls to be left in place.

Settings: prefixed all settings keys with "subliminol_"

Updated the build to provide more status output

Updated build data in project to refer to an explicit python executable.
This
will need change to be more flexible.